video22
editmodel

+++ Veterinary.Web/Controllers/PetTypesController.cs
@@ -125,21 +125,18 @@ namespace Veterinary.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(pt => pt.Pets)
+                .FirstOrDefaultAsync(pt => pt.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
-
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "The pet type can't be removed.");
+                return RedirectToAction(nameof(Index));
+            }
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -149,5 +146,6 @@ namespace Veterinary.Web.Controllers
         {
             return _context.PetTypes.Any(e => e.Id == id);
         }
+     
     }
 }
